Move methods and variables related to state file to system runner packge
